# docs/openapi/openapi.bundle.yaml
openapi: 3.0.3
info:
  title: rosmon HTTP API
  version: 1.0.0
  description: |
    REST + Server-Sent Events service di atas RouterOS API. Multi-router:
    semua endpoint (kecuali Device CRUD & Health) berada di bawah
    `/api/v1/devices/{device_id}/...` agar satu instance bisa melayani
    banyak router sekaligus.

    ## Response envelope

    Semua endpoint pakai envelope konsisten:

    - **Success single**: `{ "data": { ... } }`
    - **Success list**:   `{ "data": [ ... ], "meta": { "count": N } }`
    - **Error**:          `{ "error": { "code": "...", "message": "...", "path": "..." } }`

    Error `code` stabil — frontend bisa branch dengan aman. Mapping:

    | Code | HTTP | Trigger |
    |---|---|---|
    | `NOT_FOUND` | 404 | `mikrotik.ErrNotFound` |
    | `INVALID_ARGUMENT` | 400 | `mikrotik.ErrInvalidArgument` |
    | `AMBIGUOUS` | 409 | `mikrotik.ErrAmbiguous` |
    | `VALIDATION` | 400 | body gagal validasi `gin` binding |
    | `TIMEOUT` | 504 | `context.DeadlineExceeded` |
    | `CANCELED` | 499 | `context.Canceled` |
    | `INTERNAL` | 500 | uncategorized / panic recovery |
    | `DEVICE_NOT_FOUND` | 404 | device middleware: device ID tidak terhubung |
    | `SERVICE_UNAVAILABLE` | 503 | InfluxDB / dependency lain down |

    Setiap response menyertakan header `X-Request-ID` (16-char hex) untuk
    korelasi log.

    ## Authentication

    Service ini sekarang mendukung **JWT access + refresh token**
...

...
    4. `POST /auth/logout` revoke refresh token.

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /devices/{device_id}/reports/selling:
    parameters:
      - $ref: '#/components/parameters/DeviceIdPath'
      - $ref: '#/components/parameters/MonthQuery'
    get:
      tags:
        - Reports
      summary: List transaksi penjualan voucher
      description: |
        Query database lokal (Postgres). Tetap berfungsi meskipun router
        offline. Filter `?month=jan2026` (format mikhmon lowercase).
      responses:
        '200':
          description: List transaksi
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/TransactionResponse'
        '400':
          $ref: '#/components/responses/ErrorBadRequest'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
  /devices/{device_id}/reports/selling/today:
    parameters:
      - $ref: '#/components/parameters/DeviceIdPath'
    get:
      tags:
        - Reports
      summary: Transaksi hari ini
      description: Filter otomatis ke tanggal sekarang server (timezone server).
      responses:
        '200':
          description: Transaksi hari ini
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ReportTodayResponse'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
  /devices/{device_id}/reports/selling/summary:
    parameters:
      - $ref: '#/components/parameters/DeviceIdPath'
      - $ref: '#/components/parameters/MonthQuery'
      - $ref: '#/components/parameters/IncludeTransactionsQuery'
    get:
      tags:
        - Reports
      summary: Agregat penjualan bulanan
      description: |
        Hitung total count, total cost, total revenue, profit, dan breakdown
        per profile / per sell-price. Default current month.

        Set `?include_transactions=true` untuk include detail tiap transaksi
        di response (berat — pakai untuk export/audit).
      responses:
        '200':
          description: Summary
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ReportSummary'
              examples:
                summary_only:
                  summary: Tanpa detail transaksi
                  value:
                    data:
                      count: 142
                      total_price: 1420000
                      total_sell_price: 1704000
                      profit: 284000
                      by_profile:
                        default: 100
                        vip: 42
        '404':
          $ref: '#/components/responses/ErrorNotFound'
  /devices/{device_id}/reports/selling.csv:
    parameters:
      - $ref: '#/components/parameters/DeviceIdPath'
      - $ref: '#/components/parameters/MonthQuery'
      - $ref: '#/components/parameters/DateQuery'
    get:
      tags:
        - Reports
      summary: Export transaksi sebagai CSV
      description: |
        Download CSV file. Header: `sale_date, sale_time, username, profile,
        price, sell_price, validity, mac, ip, comment`. Filter `?month=` atau
        `?date=`.
      responses:
        '200':
          description: CSV download
          content:
            text/csv:
              schema:
                type: string
                format: binary
                example: |
                  sale_date,sale_time,username,profile,price,sell_price,validity,mac,ip,comment
                  jan/05/2026,10:30:21,vc-a1b2c3d4,default,10000,12000,168h,AA:BB:CC:DD:EE:FF,10.0.0.5,login
        '404':
          $ref: '#/components/responses/ErrorNotFound'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token (HS256). Diperoleh via `POST /auth/login`.
        Wajib di-include sebagai header `Authorization: Bearer <token>`
        kecuali untuk endpoint public (login, refresh, logout, /healthz).
  schemas:
    Envelope:
      type: object
      description: Response envelope generik. Field `data` dan `error` saling eksklusif.
      properties:
        data:
          description: Payload sukses. Tipe spesifik tergantung endpoint.
        meta:
          type: object
          additionalProperties: true
          description: Metadata opsional, biasanya berisi `count` untuk list.
          example:
            count: 47
        error:
          $ref: '#/components/schemas/ErrorPayload'
    ErrorEnvelope:
      type: object
      description: Envelope untuk response error (status 4xx/5xx).
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorPayload'
    ErrorPayload:
      type: object
      description: Detail error yang stabil — frontend bisa branch via `code`.
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Stable error code untuk branching di client.
          enum:
            - NOT_FOUND
            - INVALID_ARGUMENT
            - AMBIGUOUS
            - VALIDATION
            - TIMEOUT
            - CANCELED
            - INTERNAL
            - DEVICE_NOT_FOUND
            - SERVICE_UNAVAILABLE
        message:
          type: string
          description: Pesan error human-readable.
          example: user with id *99 not found
        path:
          type: string
          description: Path request yang menghasilkan error.
          example: /api/v1/devices/rb-main/hotspot/users/*99
        details:
          description: Payload tambahan (mis. validator detail). Tipe bebas.
...
    TransactionResponse:
      type: object
      description: |
        Satu row transaksi penjualan voucher (dicatat oleh webhook on-login
        handler untuk mode `remc`/`ntfc` saat user pertama login). Format
        `sale_date`/`sale_month` mengikuti konvensi mikhmonv3 (lowercase month).
      required:
        - id
        - sale_date
        - sale_time
        - sale_month
        - username
        - price
        - sell_price
        - created_at
      properties:
        id:
          type: integer
          format: int64
          example: 142
        sale_date:
          type: string
          example: jan/05/2026
        sale_time:
          type: string
          example: '10:30:21'
        sale_month:
          type: string
          example: jan2026
        username:
          type: string
          example: vc-a1b2c3d4
        price:
          type: integer
          example: 10000
        sell_price:
          type: integer
          example: 12000
        ip:
          type: string
          example: 10.0.0.5
        mac:
          type: string
          example: AA:BB:CC:DD:EE:FF
        validity:
          type: string
          example: 168h
        profile:
          type: string
          example: default
        comment:
          type: string
          example: login
        created_at:
          type: string
          format: date-time
          example: '2026-01-05T10:30:21+07:00'
    ReportSummary:
      type: object
      required:
        - count
        - total_price
        - total_sell_price
        - profit
      properties:
        count:
          type: integer
          example: 142
        total_price:
          type: integer
          description: Total cost (`price` × count).
          example: 1420000
        total_sell_price:
          type: integer
          description: Total revenue (`sell_price` × count).
          example: 1704000
        profit:
          type: integer
          description: total_sell_price − total_price.
          example: 284000
        by_profile:
          type: object
          additionalProperties:
            type: integer
          description: Count per profile name.
          example:
            default: 100
            vip: 42
        by_sell_price:
          type: object
          additionalProperties:
            type: integer
          description: Count per sell-price bucket (key = harga sebagai string).
          example:
            '10000': 100
            '20000': 42
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/TransactionResponse'
          description: Detail tiap transaksi. Hanya di-isi kalau query `?include_transactions=true`.
    ReportTodayResponse:
      type: object
      required:
        - date
        - count
        - transactions
      properties:
        date:
          type: string
          example: may/19/2026
        count:
          type: integer
          example: 5
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/TransactionResponse'
...
  parameters:
    DeviceIdPath:
      name: device_id
      in: path
      required: true
      description: Numeric device ID.
      schema:
        type: integer
        format: int64
        example: 1
...
    MonthQuery:
      name: month
      in: query
      required: false
      description: Filter bulan untuk laporan transaksi, format `<bln>jjjj` lowercase (mis. `jan2026`).
      schema:
        type: string
        pattern: ^(jan|feb|mar|apr|may|jun|jul|aug|sep|oct|nov|dec)[0-9]{4}$
        example: jan2026
    DateQuery:
      name: date
      in: query
      required: false
      description: Filter tanggal untuk laporan transaksi, format mikhmon `<bln>/dd/yyyy` lowercase (mis. `jan/05/2026`).
      schema:
        type: string
        example: jan/05/2026
...
    IncludeTransactionsQuery:
      name: include_transactions
      in: query
      required: false
      description: Kalau `true`, response summary mencakup detail tiap transaksi.
      schema:
        type: boolean
        default: false
...
  responses:
    ErrorBadRequest:
      description: Request tidak valid (parameter atau body salah).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            invalid_argument:
              summary: Argumen invalid
              value:
                error:
                  code: INVALID_ARGUMENT
                  message: id tidak boleh kosong
                  path: /api/v1/devices/rb-main/hotspot/users/
...
    ErrorNotFound:
      description: Resource yang dirujuk tidak ditemukan.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            user_not_found:
              summary: User tidak ada
              value:
                error:
                  code: NOT_FOUND
                  message: 'mikrotik: not found'
                  path: /api/v1/devices/rb-main/hotspot/users/*99
...